backend/auraread/urls.py
"""
Enhanced URL configuration for auraread project with AI features
"""
from django.contrib import admin
from django.urls import path, include
from django.conf import settings
from django.conf.urls.static import static
from django.views.static import serve
from django.urls import re_path
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)

urlpatterns = [
    path('admin/', admin.site.urls),
    
    # Enhanced Authentication with JWT
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    
    # App URLs
    path('api/', include('users.api.urls')),  # Auth endpoints
    path('api/', include('documents.api.urls')),
    path('api/', include('annotations.api.urls')),
    path('', include('ai_features.urls')),  # New AI features
    path('api-auth/', include('rest_framework.urls')),
    
    # Serve media files directly
    re_path(r'^media/(?P<path>.*)$', serve, {'document_root': settings.MEDIA_ROOT}),
]

# Serve media files in development
if settings.DEBUG:
    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
