backend/requirements_enhanced.txt
# Enhanced Requirements - Production Ready

# Core Django Framework
django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database & Performance
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0

# Security & CORS
django-cors-headers==4.3.1
django-ratelimit==4.1.0
python-decouple==3.8

# Enhanced PDF Processing
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# Advanced TTS Engines
gTTS==2.5.0
pyttsx3==2.90
azure-cognitiveservices-speech==1.34.0
boto3==1.34.0

# AI & Machine Learning
openai==1.3.7
sentence-transformers==2.2.2
transformers==4.35.2
torch>=2.0.0

# File Processing
python-magic==0.4.27
chardet==5.2.0

# Background Tasks (Optional)
celery[redis]==5.3.4

# Development & Debugging
django-extensions==3.2.3
